refactor: HttpRequest 객체 분리
